Always initialize copy scratch during copyinstr
The value held in copy scratch is used to determine whether or not to return an ENAMETOOLONG fault. In the case where a max length of 0 is provided to copyinstr, the loop body where the value of copy scratch is populated from the input string is skipped which could lead to a non-deterministic the return value from copyinstr.